+++ views/back/customer/src/controller/main.js
@@ -7,8 +7,8 @@
     var _ = require("underscore");
     angular.module("controller.main",[
                                         "ng.ueditor"
-    ]).controller('main',['$scope','showCtrl','dataService','FileUploader','pageResult'
-        ,function($scope,showCtrl,dataService,FileUploader,pageResult){
+    ]).controller('main',['$scope','showCtrl','dataService','FileUploader','pageResult','$window'
+        ,function($scope,showCtrl,dataService,FileUploader,pageResult,$window){
             /************************数据模型****************************/
             //设置用户的权限分配
             $scope.userPowers = [{name:"普通用户",usertype:1},{name:"名人",usertype:2}]
@@ -40,6 +40,9 @@
 
             //删除方法
             $scope.removeCustom = function(cus){
+                var tFlag = $window.confirm("真的要添加吗");
+                if(!tFlag)
+                    return false;
                 dataService.removeCustomer(cus._id)
                     .success(function(data){
                         if(data.err){
